% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getNormTwoChannel}
\alias{getNormTwoChannel}
\alias{getNormTwoChannel,Rnits-method}
\alias{getNormTwoChannel,character,ANY-method}
\title{Get Normalized channel data for two channel arrays}
\usage{
getNormTwoChannel(object)

\S4method{getNormTwoChannel}{Rnits}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Rnits}} object}
}
\value{
A list containing R and G fields for normalized Red and Green channel data respectively.
}
\description{
For two color data, extract normalized channel data from \code{\linkS4class{Rnits}} object
}

