% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{yeastchemostat}
\alias{yeastchemostat}
\title{Yeast chemostat data from Ronen and Botstein (Proc Natl Acad Sci U S A. 2006 Jan 10;103(2):389-94. Epub 2005 Dec 28.)}
\format{An \code{ExpressionSet} object with containing 'Sample' and 'Time' columns and replicates removed.}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4158
}
\usage{
yeastchemostat
}
\description{
(From author's GEO submission) Transcriptional response of
steady-state yeast cultures to transient perturbations in carbon source
}

